requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.5.0
transformers>=4.35.2
pydantic>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
huggingface_hub>=0.19.4
gunicorn>=21.2.0
redis>=5.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
